+++ modules/App/assets/vue-components/field-number.js
@@ -3,7 +3,7 @@ export default {
 
     _meta: {
         label: 'Number',
-        info: 'Quantityt etc',
+        info: 'Quantity etc',
         icon: 'system:assets/icons/number.svg',
         settings: [
             {name: 'readonly', type: 'boolean', opts: {default: false}},